@@ -18,7 +18,7 @@ classifiers =
 package_dir = 
 	= toupy
 packages = find:
-python_requires = >=3.6
+python_requires = >=3.8
 zip_safe = False
 include_package_data = True
 scripts =
