@@ -29,7 +29,7 @@ setuptools.setup(
     ],
     package_dir={"toupy": "toupy"},
     packages=setuptools.find_packages(),
-    python_requires=">=3.6",
+    python_requires=">=3.8",
     scripts=[
         "bin/file_comp",
         "bin/missing_recons",
@@ -37,18 +37,18 @@ setuptools.setup(
         "bin/create_toupy_templates",
     ],
     install_requires=[
-        "fabio>=0.11.0",
-        "h5py>=3.1.0",
-        "ipython>=7.16.1",
-        "joblib>=1.0.1",
-        "matplotlib>=3.3.4",
-        "numpy>=1.16.5",
-        "numexpr>=2.6.9",
-        "scipy>=1.5.4",
-        "PyFFTW>=0.11.1",
-        "pyopencl>=2021.1.1",
-        "scikit_image>=0.17.2",
-        "silx>=0.9.0",
-        "tqdm>=4.61.2",
+        "fabio>=2022.12.1",
+        "h5py>=3.8.0",
+        "ipython>=8.10.0",
+        "joblib>=1.2.0",
+        "matplotlib>=3.7.0",
+        "numpy>=1.24.2",
+        "numexpr>=2.8.4",
+        "scipy>=1.10.0",
+        "pyFFTW>=0.13.1",
+        "pyopencl>=2022.3.1",
+        "scikit-image>=0.19.3",
+        "silx>=1.1.2",
+        "tqdm>=4.64.1",
     ],
 )
